enforce the ui to always display the header
