+++ packages/web-app-draw-io/src/App.vue
@@ -250,11 +250,6 @@ export default {
 </script>
 <style scoped>
 #drawio-editor {
-  position: fixed;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
   width: 100%;
   height: 100%;
   border: none;
