+++ packages/web-runtime/src/layouts/Plain.vue
@@ -1,0 +1,6 @@
+<template>
+  <div>
+    <router-view />
+  </div>
+</template>
+<style lang="scss"></style>
